fix: just used middleware for protection
