--- plugins/ecs/manifest.json
@@ -1,12 +1,0 @@
-{
-  "id": "ecs",
-  "description": "ECS Task Protection",
-  "functions": [
-    {
-      "name": "Apply Task Protection",
-      "id": "protection",
-      "type": "guardrail",
-      "supportedHooks": ["beforeRequestHook", "afterRequestHook"]
-    }
-  ]
-}
