+++ src/middlewares/ecs/ECSProtectionManager.ts
