+++ src/middlewares/ecs/InMemoryProtectionManager.ts
