+++ src/middlewares/ecs/createProtectionManager.ts
