+++ src/middlewares/ecs/types.ts
